+++ src/main/kotlin/com/zepben/evolve/services/network/tracing/feeder/SetDirection.kt
@@ -8,9 +8,11 @@
 
 package com.zepben.evolve.services.network.tracing.feeder
 
+import com.zepben.evolve.cim.iec61970.base.core.ConductingEquipment
 import com.zepben.evolve.cim.iec61970.base.core.Feeder
 import com.zepben.evolve.cim.iec61970.base.core.Terminal
 import com.zepben.evolve.cim.iec61970.base.wires.PowerTransformer
+import com.zepben.evolve.cim.iec61970.base.wires.Switch
 import com.zepben.evolve.services.network.NetworkService
 import com.zepben.evolve.services.network.tracing.OpenTest
 import com.zepben.evolve.services.network.tracing.traversals.BasicTracker
@@ -55,7 +57,7 @@ class SetDirection {
      * @param network The network in which to apply feeder directions.
      */
     fun run(network: NetworkService) {
-        run(network.sequenceOf<Feeder>().mapNotNull { it.normalHeadTerminal }.toList())
+        run(network.sequenceOf<Feeder>().mapNotNull { it.normalHeadTerminal }.filter { !it.conductingEquipment.isNormallyOpenSwitch() }.toList())
     }
 
     /**
@@ -157,6 +159,9 @@ class SetDirection {
             .filter { it != terminal }
             .forEach { queue(it) }
     }
+
+    private fun ConductingEquipment?.isNormallyOpenSwitch(): Boolean =
+        (this is Switch) && isNormallyOpen()
 
     private fun BranchRecursiveTraversal<Terminal>.startNewBranch(terminal: Terminal) {
         branchQueue.add(branchSupplier().setStart(terminal))